services/emails/requirements.txt
Django
djangorestframework
djongo
psycopg2
requests
